Refactor filter components and optimize filter logic
Removed `LabeledCheckbox` from its own file and moved it inline where necessary to streamline code structure. Updated filter-related components to improve prop consistency (e.g., replacing `sxStyle` with `sx`). Enhanced filter logic by eliminating redundant operations and improving clarity in active filter handling.
