+++ packages/packages/core/editor-global-classes/src/components/search-and-filter/components/filter/clear-icon-button.tsx
@@ -4,15 +4,15 @@ import { IconButton, type SxProps, type Theme, Tooltip } from '@elementor/ui';
 
 import { useSearchAndFilters } from '../../context';
 
-type ClearIconButtonProps = { tooltipText: React.ReactNode; sxStyle?: SxProps< Theme > };
+type ClearIconButtonProps = { tooltipText: React.ReactNode; sx?: SxProps< Theme > };
 
-export const ClearIconButton = ( { tooltipText, sxStyle }: ClearIconButtonProps ) => {
+export const ClearIconButton = ( { tooltipText, sx }: ClearIconButtonProps ) => {
 	const {
 		filters: { onClearFilter },
 	} = useSearchAndFilters();
 	return (
 		<Tooltip title={ tooltipText } placement="top" disableInteractive>
-			<IconButton size="tiny" onClick={ onClearFilter } sx={ sxStyle }>
+			<IconButton size="tiny" onClick={ onClearFilter } sx={ sx }>
 				<BrushBigIcon fontSize="tiny" />
 			</IconButton>
 		</Tooltip>
